N_ptl + Dict -> List
N_ptl: before, if one added a further purpose,
the seeds for the already existing purposes and particle types changed.
Now they remain the same at the cost of
restricting the maximum number of particly types.

Dict -> Tuple: _purposes is a list, which is ordered (keeps the order),
which simplifies the tasks for the user to initilize the class.

_ptype_ind is also not needed,
it can be calculated on the fly with list(dict.keys())