store engines in 3D+mutliple realization
Engines now have a realization dimension, and init_prngs and generate
can switch back and forth between realizations.
Don't abuse this feature. This can be useful for parallelization
in some cases, but values larger than the number pipes
of the vector processor is a good sign of abuse. (tested)

clear_prng is a shortcut to initialize to empty container. (tested)

Manual and direct seed value manipulation
for debugging purposes is added. (tests needed)